requirements.txt
Jinja2==2.10
markdown2==2.3.5
MarkupSafe==1.0
python-slugify==1.2.5
Unidecode==1.0.22
